Rename package java to settings
