+++ settings/settings_darwin.go
@@ -1,4 +1,4 @@
-package java
+package settings
 
 import "path/filepath"
