+++ settings/settings_linux.go
@@ -1,4 +1,4 @@
-package java
+package settings
 
 import "path/filepath"
